feat: gene page refactor
