+++ src/components/ExternalLink/index.js
@@ -1,0 +1,2 @@
+export { default as ExternalLink } from './ExternalLink';
+export { default as XRefLinks } from './XRefLinks';
