+++ src/components/Header/index.js
@@ -1,0 +1,2 @@
+export { default } from './Header';
+export { default as OTButtonLink } from './OTButtonLink';
